Add scenario to the model

Add scenario to the model

Usage

add_scenario(
  model,
  name = "expansion",
  origin = "baseline",
  origin_start = NA,
  origin_end = NA,
  shock
)

Arguments

model

SFC model object

name

string name of scenario, defaults to 'expansion'

origin

string name of origin scenario, from which the new scenario will be created, defaults to 'baseline'

origin_start

numeric period number from origin scenario from which the new scenario will begin

origin_end

numeric period number from origin scenario on which the new scenario will end

shock

shock object from create_shock() and add_shock()

Value

updated SFC model object containing added scenario