Network plot of the model

Network plot of the model

Usage

plot_cycles(model, save_file = NULL)

Arguments

model

SFC model object created with create_model

save_file

name and path to save the plot as html file

Value

visNetwork object

Details

This function creates a representation of a model as a directed graph. Additionally it shows cycles in the model including these with lagged variables. Graph can be saved as html file.

Examples

model <- godley::create_model(name = "SFC model", template = "BMW")
#> Model BMW loaded from template
plot_cycles(model)
#> Model prepared successfully