Make initial matrix row for baseline scenario and prepare equations for simulate_scenario()

Make initial matrix row for baseline scenario and prepare equations for simulate_scenario()

Usage

prepare(model, verbose = FALSE)

Arguments

model

SFC model object

verbose

logical to tell if additional model verbose should be displayed

Value

verified and prepared SFC model object