Simulate scenario of SFC model object

Simulate scenario of SFC model object

Usage

simulate_scenario(
  model,
  scenario,
  periods = NA,
  start_date = NA,
  method = "Gauss",
  max_iter = 350,
  tol = 1e-05,
  hidden_tol = 0.1,
  verbose = FALSE
)

Arguments

model

SFC model object

scenario

vector of strings or single string name of scenario(s) to simulate

periods

numeric total number of rows (periods) in the model, defaults to 100

start_date

character date to begin the simulation in the format "yyyy-mm-dd"

method

string name of method used to find solution chosen from: 'Gauss', 'Newton', defaults to 'Gauss'

max_iter

numeric maximum iterations allowed per period, defaults to 350

tol

numeric tolerance accepted to determine convergence, defaults to 1e-05

hidden_tol

numeric error tolerance to accept the equality of hidden equations, defaults to 0.1.

verbose

logical to tell if additional model verbose should be displayed

Value

updated model containing simulated scenario(s)